Oops forgot the rest of the gcc3.4 patch

+++ licq/src/user.cpp
@@ -3578,7 +3578,7 @@ ICQOwner::ICQOwner()
   m_fConf.ReadBool("HideIP", m_bHideIp, false);
   m_fConf.ReadNum("RCG", m_nRandomChatGroup, ICQ_RANDOMxCHATxGROUP_NONE);
   m_fConf.ReadStr("AutoResponse", szTemp, "");
-  m_fConf.ReadNum("SSTime", (unsigned long)m_nSSTime, 0L);
+  m_fConf.ReadNum("SSTime", m_nSSTime, 0L);
   m_fConf.ReadNum("SSCount", m_nSSCount, 0);
   SetAutoResponse(szTemp);
 
@@ -3647,7 +3647,7 @@ ICQOwner::ICQOwner(const char *_szId, unsigned long _nPPID)
   m_fConf.ReadBool("HideIP", m_bHideIp, false);
   m_fConf.ReadNum("RCG", m_nRandomChatGroup, ICQ_RANDOMxCHATxGROUP_NONE);
   m_fConf.ReadStr("AutoResponse", szTemp, "");
-  m_fConf.ReadNum("SSTime", (unsigned long)m_nSSTime, 0L);
+  m_fConf.ReadNum("SSTime", m_nSSTime, 0L);
   m_fConf.ReadNum("SSCount", m_nSSCount, 0);
   SetAutoResponse(szTemp);
 
@@ -3681,7 +3681,7 @@ ICQOwner::~ICQOwner()
   }
   m_fConf.SetSection("user");
   m_fConf.WriteStr("AutoResponse", AutoResponse());
-  m_fConf.WriteNum("SSTime", (unsigned long)m_nSSTime);
+  m_fConf.WriteNum("SSTime", m_nSSTime);
   m_fConf.WriteNum("SSCount", m_nSSCount);
   if (!m_fConf.FlushFile())
   {
@@ -3735,7 +3735,7 @@ void ICQOwner::SaveLicqInfo()
   m_fConf.WriteBool("HideIP", HideIp());
   m_fConf.WriteBool("Authorization", GetAuthorization());
   m_fConf.WriteNum("RCG", RandomChatGroup());
-  m_fConf.WriteNum("SSTime", (unsigned long)m_nSSTime);
+  m_fConf.WriteNum("SSTime", m_nSSTime);
   m_fConf.WriteNum("SSCount", m_nSSCount);
   
   if (m_bSavePassword)
